+++ handin_1/src/utils.cpp
@@ -74,7 +74,7 @@ namespace Utils
         return pow(2.0, hashbites);
     }
 
-    int hashBitsToIdx(u_char hash, int hashbits)
+    long long hashBitsToIdx(u_char hash, int hashbits)
     {
         long long mask = ((1ULL << (hashbits & 0x3F)) & -(hashbits != 64)) - 1;
         return hash & mask;
